s:GetSecondSearchAction(): Handle invalid offsets
Re-parse with more lenient regular expression as a fallback, to still consume the pattern even if followed by an invalid offset.
Use intermediate l:parse variable instead of relying on Vim exception; this was ugly, anyway.
Check for pathological empty a:offset to avoid returning an invalid "let dummy =" command.